First commit (work in progress)
To be shared with my mentor Prasad at the Udacity's Self-Driving
Nanodegree (Advanced Lane Finding module).

Please go over the IPython notebooks to get a glimpse of my thought
process in developing this code and learning how to build a software
stack targetting the task of lane finding. Great course.